feat: update test

+++ contracts/MapleTreasury.sol
@@ -112,7 +112,7 @@ contract MapleTreasury {
     */
     function passThroughFundsToken() isGovernor public {
         IERC20 _fundsToken = IERC20(fundsToken);
-        uint passThroughAmount = _fundsToken.balanceOf(address(this));
+        uint256 passThroughAmount = _fundsToken.balanceOf(address(this));
         require(_fundsToken.transfer(mpl, passThroughAmount), "MapleTreasury:FUNDS_RECEIVE_TRANSFER");
         IMapleToken(mpl).updateFundsReceived();
         emit PassThrough(msg.sender, passThroughAmount);
